task4_automate_all/requirements.txt
mysql-connector-python==8.0.26
dask==2021.8.1
SQLAlchemy==1.4.23